display rating component + notification display
