+++ frontend/src/components/components.js
@@ -42,6 +42,7 @@ import renderDatePicker from "./fields/renderDatePicker";
 import renderDateSimplePicker from "./fields/renderDateSimplePicker";
 import UserAutocomplete from "./fields/UserAutocomplete";
 import RestaurauntAutocomplete from "./fields/RestaurauntAutocomplete";
+import Prices from './fields/Prices'
 
 //Map
 import SearchMap from "./map/SearchMap";
@@ -60,6 +61,7 @@ import ProgressIcon from "./general/ProgressIcon";
 import GroupAvatars from "./general/GroupAvatars";
 import NotificationWrapper from "./general/NotificationWrapper";
 import Calendar from "./general/Calendar";
+import DisplayRating from './general/DisplayRating'
 
 //User
 import Settings from "./user/Settings";
@@ -87,6 +89,9 @@ import Comment from "./restaurant/Comment";
 //Search
 import SearchPage from './search/SearchPage';
 import SearchBar from "./search/SearchBar";
+
+//Skeletons
+import SkeletonRestaurant from './skeleton/SkeletonRestaurant'
 
 export {
   ErrorPage,
@@ -156,5 +161,8 @@ export {
   NotificationWrapper,
   Calendar,
   ProgressIcon,
-  SearchPage
+  SearchPage,
+  SkeletonRestaurant,
+  Prices,
+  DisplayRating
 };
