+++ frontend/src/styles/profile.module.css
@@ -61,30 +61,62 @@
   display: flex;
   width: 90%;
   margin: 0 auto;
-  justify-content: center;
-  flex-direction: column;
   align-items: center;
+  flex-direction: column; 
 }
 
 .activity {
   display: flex;
-  padding: .75rem 1rem;
+  flex-direction: column;
   background-color: white;
-  box-shadow: var(--shadow-0);
-  margin: .75rem 0;
+  border: var(--border-separator);
+  margin: .75rem .25rem;
   justify-content: space-between;
   font-size: 0.85rem;
-  flex: 1;
-  width: 100%;
-  height: 60px;
+  width: 350px;
+  border-radius: 5px;
+}
+
+.activity:hover{
+  box-shadow: var(--shadow-1);
+}
+
+.activity:hover .activitydate{
+  cursor: pointer;
+}
+
+.activityContent {
+  border-bottom: var(--border-separator);
+  word-wrap: break-word;
+  padding: 1rem 1rem;
+  font-weight: 500;
+  font-size: .8rem;
 }
 
 .activitydate {
   display: flex;
+  justify-content: flex-end;
   align-items: center;
   white-space: nowrap;
+  font-weight: 500;
+  font-family: "Roboto";
   font-size: 0.7rem;
   color: #696969;
+  padding:.5rem;
+}
+
+.infoAll {
+  display: flex;
+  align-items: center;
+  margin-left: 5px;
+  justify-content: space-between;
+  flex: 1;
+}
+
+.infoActivity{
+  color: var(--grey-text);
+  font-size: .7rem;
+  font-weight: 500;
 }
 
 .activityLog {
@@ -97,7 +129,8 @@
 
 .activityinfo {
   display: flex;
-  align-items: center;
+  padding: .5rem .75rem;
+  border-bottom: var(--border-separator);
 }
 
 .activity:last-child {
@@ -105,13 +138,30 @@
 }
 
 .useravatar {
-  transform: scale(0.8);
-}
-
-.link {
-  color: red;
+  transform: scale(0.9);
 }
 
 .space {
   margin-left: 3px;
+}
+
+
+.imageWrapper{
+  width: 100%;
+  height: 200px; /* or other desired height */
+  overflow: hidden;
+  position: relative;
+  box-shadow: var(--shadow-0);
+  border-top-right-radius: 5px;
+  border-top-left-radius: 5px;
+}
+
+.image{
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  height: 100%;
+  width: 100%;
+  background-size: cover;
+  background-position: 50%;
 }
